complited todo API

+++ app/Http/Controllers/TaskDestroyController.php
@@ -8,8 +8,16 @@ use Illuminate\Http\Request;
 
 class TaskDestroyController extends Controller
 {
-    public function __invoke(Task $task): JsonResponse
+    public function __invoke($id): JsonResponse
     {
+        $task = Task::find($id);
+
+        if(!$task){
+            return response()->json([
+                'message' => 'not found'
+            ], 404);
+        }
+
         $task->delete();
 
         return response()->json([
